refactor : page네이밍 수정, index.ts에서 일괄 export
