+++ src/pages/index.ts
@@ -1,0 +1,5 @@
+export { default as Home } from './Home';
+export { default as Login } from './Login';
+export { default as My } from './My';
+export { default as NotFound } from './NotFound';
+export { default as Order } from './Order';
